+++ stack.yaml.lock
@@ -5,29 +5,16 @@
 
 packages:
 - completed:
-    size: 288620
-    url: https://github.com/haskoin/haskoin-core/archive/6c3fdcf771fc92e7e5c2c7da62d6988447d0c4f6.tar.gz
+    size: 445088
+    url: https://github.com/haskoin/haskoin-core/archive/1196c5a3efffc887670c62c6ead2ec45d9d3a5db.tar.gz
     name: haskoin-core
-    version: 0.14.1
-    sha256: 787a8afc0f6fe773e2f7a4fa75904785ac6a6692a4b89224cc73306f3beb0023
+    version: 0.20.0
+    sha256: d711eb32e44d6ac70a0f65ceb0a33df1f65c460d6ca282176fec2f24a0347c0f
     pantry-tree:
-      size: 5370
-      sha256: c8a377f7050cd636fa211a813999a885249c17d64b25230099bf8a9d0c0cbc00
+      size: 6319
+      sha256: a865c7e7aeefc728ccd5d53cba7e850ffda1aad11fdd9fad8831dd3858997710
   original:
-    url: https://github.com/haskoin/haskoin-core/archive/6c3fdcf771fc92e7e5c2c7da62d6988447d0c4f6.tar.gz
-- completed:
-    size: 94640
-    subdir: data
-    url: https://github.com/haskoin/haskoin-store/archive/77fdf2170b17674221d3f782def14217ca0f4a5a.tar.gz
-    name: haskoin-store-data
-    version: 0.32.2
-    sha256: e1faaf6b4d75f39a9879379fe87a8472138b0f864c76c3fb13d05975e79b97d3
-    pantry-tree:
-      size: 623
-      sha256: 4259a7bc5832570aeb78b2cca7674d9bcd33d6d0bb67d2134f5495a723ae6637
-  original:
-    subdir: data
-    url: https://github.com/haskoin/haskoin-store/archive/77fdf2170b17674221d3f782def14217ca0f4a5a.tar.gz
+    url: https://github.com/haskoin/haskoin-core/archive/1196c5a3efffc887670c62c6ead2ec45d9d3a5db.tar.gz
 - completed:
     hackage: murmur3-1.0.4@sha256:a0038867cb80b5bb3e400b57da7af501c3e2b5583792eb7ed6d69862817d118c,2002
     pantry-tree:
@@ -50,15 +37,15 @@ packages:
   original:
     hackage: parsec-extra-0.2.0.0
 - completed:
-    hackage: secp256k1-haskell-0.3.0@sha256:0dc91562e4fa5b6090ad6025a73920058190776fb714980f44a9c7c4513f60f1,1900
+    hackage: secp256k1-haskell-0.5.0@sha256:26d8897b0b81a1ca46134e3b1cc7784c975880e2334936d92681b07ae9703c26,2158
     pantry-tree:
-      size: 600
-      sha256: 11f66cd618c45519b1bd88470df0482ec10d8732d4abc6b159ea1cc5c67cc22f
+      size: 599
+      sha256: e1e0b1d98213c7874a4260b946d1024571ac659e81822d51ea1df7d00f4c14b0
   original:
-    hackage: secp256k1-haskell-0.3.0
+    hackage: secp256k1-haskell-0.5.0
 snapshots:
 - completed:
-    size: 531237
-    url: https://raw.githubusercontent.com/commercialhaskell/stackage-snapshots/master/lts/16/0.yaml
-    sha256: 210e15b7043e2783115afe16b0d54914b1611cdaa73f3ca3ca7f8e0847ff54e5
-  original: lts-16.0
+    size: 563103
+    url: https://raw.githubusercontent.com/commercialhaskell/stackage-snapshots/master/lts/17/4.yaml
+    sha256: f11e2153044f5f71ea7b1c9398f4721f517c9bd37642ed769647b896564021f3
+  original: lts-17.4
